% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_infection_proba.R
\name{add_infection_proba}
\alias{add_infection_proba}
\title{Add/change infection probabilities in a ctdata object}
\usage{
add_infection_proba(x, proba)
}
\arguments{
\item{x}{a \code{ctdata} object}

\item{proba}{a named \code{list} of probabilities for each exposure type}
}
\value{
a \code{ctdata} object with an updated \code{risk} table
}
\description{
This function can be used to add or change infection probabilities in a
\code{ctdata} object. It is useful for changing the probabilities of infection for
different types of exposures. Probabilities are provided as a named \code{list}
which must have one probability for each exposure type in the \code{ctdata}
object.
}
\author{
Thibaut Jombart
}
