% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_infected.R
\name{add_p_infected}
\alias{add_p_infected}
\title{Attach probabilities of infections to a ctdata linelist}
\usage{
add_p_infected(x, p_inf)
}
\arguments{
\item{x}{the \code{ctdata} object to which the scores should be added}

\item{p_inf}{a named numeric vector of probabilities of infeciton as returned
by \code{\link[=p_infected]{p_infected()}}}
}
\value{
The \code{ctdata} object \code{x}, with a \code{score} column added to its
\code{linelist}.
}
\description{
Adds the per-contact probabilities of infections returned by \code{\link[=p_infected]{p_infected()}}
as a \code{p_infected} column of the \code{ctdata}'s \code{linelist}, matched by
}
\seealso{
\code{\link[=p_infected]{p_infected()}} to calculate the individual probabilities of
infection, and \code{\link[=ctscore]{ctscore()}} and \code{\link[=add_ctscore]{add_ctscore()}} for followup scores.
}
