% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p_infection.R
\name{calculate_p_infection}
\alias{calculate_p_infection}
\title{Calculate the probability of infection from past exposures}
\usage{
calculate_p_infection(x)
}
\arguments{
\item{x}{a vector of probabilities of infection for different successive
exposures, ordered by increasing date}
}
\description{
Internal. Used to calculate the probability that an infection took place at
a given exposure, accounting for previous ones, for a single contact. This
function returns a vector of probabilities of infection for different
successive exposures, ordered by increasing date. To calculate the overal
probability that a contact has been infected given a series of exposures,
see the user-facing function \code{\link[=p_infected]{p_infected()}}.
}
\author{
Thibaut Jombart
}
