% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_indiv_data.R
\name{extract_indiv_data}
\alias{extract_indiv_data}
\title{Extract individual data from a \code{ctdata} object}
\usage{
extract_indiv_data(x)
}
\arguments{
\item{x}{a \code{ctdata} object as returned by \code{\link[=make_ctdata]{make_ctdata()}} or \code{\link[=sim_ctdata]{sim_ctdata()}}}
}
\description{
\code{ctdata} objects are used to store exposure data; as contacts may report
multiple exposures, the corresponding individual data (e.g. location, data of
last visit) are repeated. This function extracts individual data so that the
output has a single row for each individual, effectively stripping exposure
data.
}
\examples{
## with a simple ctdata
x <- make_ctdata(
  contact_id = c(1, 1, 2, 3, 3), 
  date = Sys.Date() - c(6, 4, 2, 2, 3),
  type = c("normal", "funeral", "normal", "normal", "funeral"),
  location = c("some-town", "some-town", "some-town", "sincity", "sincity"),
  infection_proba = list(normal = 0.2, funeral = 0.9),
  last_visit = Sys.Date() - c(4, 4, 1, NA, NA)
)
x

extract_indiv_data(x)

## with a ctdata object containing scores (returned by ctscore())
scores <- ctscore(x, 
  incub = c(0, 0, 1, 2, 4, 3, 2, 1), 
  current_date = Sys.Date(), 
  out_type = "ctdata"
)

scores
extract_indiv_data(scores)

}
\author{
Thibaut Jombart
}
