% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_infected.R
\name{p_infected}
\alias{p_infected}
\title{Calculate the probability of infection}
\usage{
p_infected(x)
}
\arguments{
\item{x}{a \code{ctdata} object as returned by \code{\link[=make_ctdata]{make_ctdata()}}}
}
\value{
A named numeric vector giving the probability of infection for each
contact, named by contact ID. Use \code{\link[=add_p_infected]{add_p_infected()}} to attach these results
to the \code{linelist} of the source \code{ctdata}.
}
\description{
This function calculates the probability of a contact having been infected
given prior exposures, as described in Jombart et al. 2026. Unlike
\code{ctscore()}, it requires only exposure data, and no information on follow-up
history or incubation period.
}
\examples{
## create ctdata
x <- make_ctdata(
  exposures = tibble::tibble(
    contact_id = c(1, 1, 2, 3, 4),
    date = Sys.Date() - c(6, 4, 5, 1, 5),
    exposure_type = c("normal", "funeral", "normal", "normal", "null")
  ),
  linelist = tibble::tibble(
    contact_id = c(1, 2, 3, 4),
    last_visit_date = Sys.Date() - c(2, 1, 1, 3)
  ),
  infection_proba = list(normal = 0.2, funeral = 0.9, null = 0)
)

## a named vector of proba of infection
p_inf <- p_infected(x)
p_inf

## attach the scores to the ctdata linelist
x <- add_p_infected(x, p_inf)
x

}
\seealso{
\code{\link[=add_p_infected]{add_p_infected()}} to attach the results back onto the \code{ctdata}.
}
\author{
Thibaut Jombart
}
