% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctdata.R
\name{plot.ctdata}
\alias{plot.ctdata}
\title{Plot a contact-tracing timeline}
\usage{
\method{plot}{ctdata}(x, ...)
}
\arguments{
\item{x}{a \code{ctdata} object returned by \code{\link[=make_ctdata]{make_ctdata()}}, or a \code{sim_ctdata}
object returned by \code{\link[=sim_ctdata]{sim_ctdata()}}.}

\item{...}{currently ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces a timeline plot of contact-tracing data, with one row per contact
along a time axis. Contacts are ordered by first exposure day.
}
\examples{
\dontrun{
## simulated data (also shows symptom onset)
plot(sim_ctdata())

}
}
\author{
Cyril Geismar
}
