% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdata.R
\name{print.ctdata}
\alias{print.ctdata}
\title{Print a \code{ctdata} object}
\usage{
\method{print}{ctdata}(x, ...)
}
\arguments{
\item{x}{a \code{ctdata} object}

\item{...}{passed to the print methods of the underlying tibbles}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{ctdata} object
}
