% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctscore.R
\name{ctscore}
\alias{ctscore}
\title{Calculate contact tracing score}
\usage{
ctscore(x, incub)
}
\arguments{
\item{x}{a \code{ctdata} object as returned by \code{\link[=make_ctdata]{make_ctdata()}}}

\item{incub}{the incubation period distribution; can be provided as a vector
of probabilities giving p(0 day), p(1 day), p(2 days) ... or as a
\code{distcrete} object as returned by distcrete::distcrete()}
}
\description{
This function calculates the probability of detecting symptoms in a contact
on a given day given their exposure and follow-up history, as described in
Jombart et al. 2026.
}
\author{
Thibaut Jombart
}
