% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctscore.R
\name{ctscore}
\alias{ctscore}
\title{Calculate contact tracing score}
\usage{
ctscore(x, incub, current_date = Sys.Date())
}
\arguments{
\item{x}{a \code{ctdata} object as returned by \code{\link[=make_ctdata]{make_ctdata()}}}

\item{incub}{the incubation period distribution; can be provided as a vector
of probabilities giving p(0 day), p(1 day), p(2 days) ... or as a
\code{distcrete} object as returned by distcrete::distcrete()}

\item{current_date}{the current date, provided either as a \code{numeric} value or
as a \code{Date}; defaults to the current date as returned by \code{Sys.Date()}}
}
\value{
A named numeric vector giving the probability of detecting symptoms
for each contact, named by contact ID. Use \code{\link[=add_ctscore]{add_ctscore()}} to attach these
scores to the \code{linelist} of the source \code{ctdata}.
}
\description{
This function calculates the probability of detecting symptoms in a contact
on a given day given their exposure and follow-up history, as described in
Jombart et al. 2026.
}
\examples{
x <- make_ctdata(
  exposures = tibble::tibble(
    contact_id = c(1, 1, 2, 3, 4),
    date       = Sys.Date() - c(6, 4, 5, 1, 5),
    type       = c("normal", "funeral", "normal", "normal", "null")
  ),
  linelist = tibble::tibble(
    contact_id = c(1, 2, 3, 4),
    last_visit = Sys.Date() - c(2, 1, 1, 3)
  ),
  infection_proba = list(normal = 0.2, funeral = 0.9, null = 0)
)

## incubation time PMF from day 0 to 7
incub <- c(0, 0, 1, 2, 4, 3, 2, 1)

## a named vector of scores
score <- ctscore(x, incub)
score

## attach the scores to the ctdata linelist
add_ctscore(x, score)

## incubation as a distcrete object
incub <- distcrete::distcrete("gamma", interval = 1, shape = 2, scale = 2.5, w = 0)
ctscore(x, incub)
}
\seealso{
\code{\link[=add_ctscore]{add_ctscore()}} to attach the scores back onto the \code{ctdata}.
}
\author{
Thibaut Jombart
}
