% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ctdata.R
\name{make_ctdata}
\alias{make_ctdata}
\title{Build a ctdata object}
\usage{
make_ctdata(contact_id, date, type = "default", location = "default")
}
\arguments{
\item{contact_id}{a \code{character} or a \code{numeric} vector indicating identifiers
for the contacts}

\item{date}{a \code{Date}, \code{numeric}, or \code{character} vector indicating dates of
exposures; \code{character} will be converted to \code{Date} using \code{as.Date}, with
expected formats "\%Y-\%m-\%d" or "\%Y/\%m/\%d"; fancier conversions should be
done before creating a \code{ctdata} object}

\item{type}{a \code{character} used to describe the type of exposure; defaults to
\code{default}}

\item{location}{a \code{character} used to describe the geographic location of the
contact; defaults to \code{default}}
}
\description{
This constructor will build a \code{ctdata} object from different inputs
describing past exposures and follow-up history for different individuals.
}
\author{
Thibaut Jombart
}
