% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_followup.R
\name{sim_followup}
\alias{sim_followup}
\title{Simulate follow-up for contact tracing data}
\usage{
sim_followup(
  x,
  time = 1,
  delay = 1,
  duration = 21,
  coverage = 0,
  strategy = c("random", "geo_random", "ctscore", "geo_ctscore")
)
}
\arguments{
\item{x}{a \code{ctdata} object}

\item{time}{an \code{integer} indicating the number of days to run the simulation
for; starts from the most recent date in \code{x}, be it in follow-up
history or in exposures; defaults to 1 day}

\item{delay}{an \code{integer} the minimum delay for follow-up to start, after the
first exposure of the concerned contact; defaults to 1 - visit can start
the day after the first exposure}

\item{duration}{an \code{integer} indicating the number of days after the last
exposure a contact should be followed for; usually determined according to
the incubation time distribution; defaults to 21 days}

\item{coverage}{the expected proportion of contacts visited at any time step; defaults
to 0 - no follow-up}

\item{strategy}{a \code{character} indicating the follow-up strategy to use in the
simulations; currently available values are: "random"; see details section
for more information}
}
\value{
\code{x} with columns \code{last_visit} and \code{detection_date} updated according to the follow-up strategy.
}
\description{
This function adds simulated follow-up to a \code{ctdata} object as generated by
\code{\link[=make_ctdata]{make_ctdata()}} or simulated via \code{\link[=sim_ctdata]{sim_ctdata()}}. The function simulates
follow-up for each individual in the \code{ctdata} object, based on the specified
follow-up time and the probability of being followed up. The function returns
a modified \code{ctdata} object with the simulated follow-up data added.
}
\details{
Available follow-up strategies (\code{strategy} argument) include:
\itemize{
\item "random": individuals are visited at random every day of the simulation
\item "geo_random": locations are prioritized at random every day of the
simulation; as many contacts as possible are visited in the first chosen
location, then if capacity remains, in the second, third, etc.
\item "ctscore": individuals are prioritized by highest ctscore every day of the
simulation
\item "geo_ctscore": geographic locations are prioritized according by highest
ctscore every day of the simulation; as many contacts as possible are
visited in the first chosen location, then if capacity remains, in the
second, third, etc.
}
}
\examples{
sim_ctdata() |> 
sim_followup(coverage = 0.2, delay = 7, strategy = "random") |> 
plot()

}
\seealso{
\code{\link[=make_ctdata]{make_ctdata()}} to create contact tracing data from existing data,
or \code{\link[=sim_ctdata]{sim_ctdata()}} to simulate them
}
\author{
Thibaut Jombart
}
